+++ src/frontend/apps/impress/src/cunningham/__tests__/useCunninghamTheme.spec.tsx
@@ -1,9 +1,9 @@
 import { useCunninghamTheme } from '../useCunninghamTheme';
 
 describe('<useCunninghamTheme />', () => {
-  it('has the dsfr logo correctly set', () => {
+  it('has the logo correctly set', () => {
     const { themeTokens, setTheme } = useCunninghamTheme.getState();
-    setTheme('dsfr');
+    setTheme('default');
     const logo = themeTokens().logo;
     expect(logo?.src).toBe('/assets/logo-gouv.svg');
     expect(logo?.widthHeader).toBe('110px');